pipelines/file_sync/requirements.txt
google-api-python-client>=2.0
google-auth-httplib2>=0.1
google-auth-oauthlib>=1.0
PyGithub>=2.0
python-dotenv>=1.0
colorama>=0.4
tqdm>=4.0
